+++ application-client/src/main/kotlin/org/fastcampus/applicationclient/cart/controller/CartController.kt
@@ -1,14 +1,19 @@
 package org.fastcampus.applicationclient.cart.controller
 
 import org.fastcampus.applicationclient.cart.controller.dto.request.CartMenuInsertionRequest
+import org.fastcampus.applicationclient.cart.controller.dto.request.CartUpdateRequest
 import org.fastcampus.applicationclient.cart.controller.dto.response.CartResponse
+import org.fastcampus.applicationclient.cart.controller.dto.response.CartUpdateResponse
 import org.fastcampus.applicationclient.cart.service.CartMenuInsertionService
+import org.fastcampus.applicationclient.cart.service.CartReadService
+import org.fastcampus.applicationclient.cart.service.CartUpdateService
 import org.fastcampus.applicationclient.config.security.dto.AuthMember
 import org.fastcampus.applicationclient.config.security.dto.JwtAuthenticated
 import org.fastcampus.common.dto.APIResponseDTO
 import org.springframework.http.HttpStatus
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -18,6 +23,8 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/api/v1/carts")
 class CartController(
     private val cartMenuInsertionService: CartMenuInsertionService,
+    private val cartReadService: CartReadService,
+    private val cartUpdateService: CartUpdateService,
 ) {
     @JwtAuthenticated
     @PostMapping
@@ -34,7 +41,17 @@ class CartController(
     fun getCart(
         @AuthenticationPrincipal authMember: AuthMember,
     ): APIResponseDTO<CartResponse> {
-        val cartResponse = cartMenuInsertionService.getCart(authMember.id)
+        val cartResponse = cartReadService.getCart(authMember.id)
         return APIResponseDTO(HttpStatus.OK.value(), HttpStatus.OK.reasonPhrase, cartResponse)
+    }
+
+    @JwtAuthenticated
+    @PatchMapping
+    fun updateCart(
+        @AuthenticationPrincipal authMember: AuthMember,
+        @RequestBody request: CartUpdateRequest,
+    ): APIResponseDTO<CartUpdateResponse> {
+        val response = cartUpdateService.updateCart(authMember.id, request)
+        return APIResponseDTO(HttpStatus.OK.value(), HttpStatus.OK.reasonPhrase, response)
     }
 }
