+++ application-client/src/main/kotlin/org/fastcampus/applicationclient/cart/controller/dto/request/CartUpdateRequest.kt
@@ -1,0 +1,6 @@
+package org.fastcampus.applicationclient.cart.controller.dto.request
+
+data class CartUpdateRequest(
+    val cartId: Long,
+    val quantity: Long,
+)
