+++ application-client/src/main/kotlin/org/fastcampus/applicationclient/cart/controller/dto/response/CartUpdateResponse.kt
@@ -1,0 +1,6 @@
+package org.fastcampus.applicationclient.cart.controller.dto.response
+
+data class CartUpdateResponse(
+    val cartId: Long,
+    val quantity: Long,
+)
